requirements.txt
# Data processing
pyspark>=3.5.0

# API requests
requests>=2.31

# AWS S3 interaction
boto3>=1.34
s3fs>=2024.3.1

# Orchestration
apache-airflow==2.9.0
# apache-airflow-providers-apache-spark>=4.3.1
# apache-airflow-providers-amazon>=8.11.0

# Testing
pytest>=8.0
pytest-mock>=3.14

# Extras
# python-dotenv>=1.0
